Moved MultiFrame Feature Extraction and adapted the model to use it
Moved MultiFrame Feature Extraction and adapted the model to use it.
